fix(migration): 修复模型Device, IP等声明定义

+++ database/migrations/2019_07_25_022836_create_devices_table.php
@@ -7,19 +7,19 @@ use Illuminate\Database\Migrations\Migration;
 
 class CreateDevicesTable extends Migration
 {
-    public function up()
-    {
-        Schema::create('devices', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('key')->nullable()->unique();
-            $table->string('user_agent');
-            $table->timestamp('created_at');
-        });
-    }
+  public function up()
+  {
+    Schema::create('devices', function (Blueprint $table) {
+      $table->increments('id');
+      $table->string('key')->nullable()->unique();
+      $table->string('user_agent');
+      $table->timestamp('created_at');
+    });
+  }
 
-    public function down()
-    {
-        Schema::dropIfExists('devices');
-        Series::delete('device_id');
-    }
+  public function down()
+  {
+    Schema::dropIfExists('devices');
+    Series::delete('device_id');
+  }
 }
