+++ database/migrations/2019_07_25_022854_create_ip_table.php
@@ -6,19 +6,19 @@ use Illuminate\Database\Migrations\Migration;
 
 class CreateIpTable extends Migration
 {
-    public function up()
-    {
-        Schema::create('ip', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('address');
-            $table->string('province');
-            $table->string('city');
-            $table->timestamp('updated_at');
-        });
-    }
+  public function up()
+  {
+    Schema::create('ip', function (Blueprint $table) {
+      $table->increments('id');
+      $table->string('address');
+      $table->string('province');
+      $table->string('city');
+      $table->timestamp('updated_at');
+    });
+  }
 
-    public function down()
-    {
-        Schema::dropIfExists('ip');
-    }
+  public function down()
+  {
+    Schema::dropIfExists('ip');
+  }
 }
